Refactor simulation time handling and add unit conversion: Change simulationTimeSeconds method to return double for more precise time representation; introduce UnitConverter class for converting between MPH and MPS.

+++ src/main/java/trafficsim/core/sim/SimulationEngine.java
@@ -96,8 +96,8 @@ public final class SimulationEngine
         }
     }
 
-    public Long simulationTimeSeconds()
+    public double simulationTimeSeconds()
     {
-        return simTimeMillis.get() / 1000;
+        return simTimeMillis.get() / 1000.0;
     }
 }
